requirements.txt
# Flask 및 웹 프레임워크
Flask>=2.3.0
Flask-CORS>=4.0.0

# 데이터베이스
# sqlite3  # Python 내장 모듈이므로 제거

# JWT 인증
PyJWT>=2.8.0
cryptography>=41.0.0

# LangGraph RAG 시스템
langgraph>=0.2.0
chromadb>=0.4.0
numpy>=1.24.0
requests>=2.31.0

# 이미지 처리
Pillow>=10.0.0

# Clova OCR (실제 패키지명은 다를 수 있음)
# clova-ocr-sdk>=1.0.0

# WSGI 서버 (프로덕션용)
gunicorn>=21.0.0

# 기타 유틸리티
python-dotenv>=1.0.0
